+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to `period` will be documented in this file
 ## 2.0.0
 
 - Breaking Change: `Period::length()` now uses the Period's precision instead of always returning days
+- `Period::duration()` returns an instance of `Duration`
+- `PeriodCollection::duration()` returns an instance of `Duration` 
+  with the sum of all included Period's durations
 
 ## 1.1.1 - 2019-02-01